requirements-frontend.txt
# Enhanced Frontend Requirements for GENESIS-AI
# Additional packages for advanced visualization and features

# Enhanced Streamlit components
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-plotly-events>=0.0.6
streamlit-aggrid>=0.3.4
streamlit-elements>=0.1.0

# Advanced visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
bokeh>=3.2.0
altair>=5.1.0

# 3D visualization and scientific computing
vtk>=9.2.0
pyvista>=0.42.0
mayavi>=4.8.1

# Machine learning and anomaly detection
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0
pyod>=1.1.0

# Image processing for explainability
Pillow>=10.0.0
opencv-python>=4.8.0

# Performance monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# Enhanced data processing
polars>=0.19.0
dask>=2023.9.0

# Geospatial visualization
folium>=0.14.0
geopandas>=0.14.0
cartopy>=0.22.0

# Real-time features
websockets>=11.0.0
asyncio-mqtt>=0.13.0

# Audio feedback (optional)
pygame>=2.5.0
pydub>=0.25.0

# Enhanced styling
jinja2>=3.1.0
markdown>=3.5.0
